Made it include Everything and improved code
Some AJAX stuff still won't be able to come. CORS needs to be disabled using the Moesif chrome extension for local testing. Once the server is uploaded to a permanent IP, this can change. Also, need to adjust the java server to be able to handle asynchronous requests before we can use this extension (too slow otherwise)